Added splash page and background image
A splash page (basically one extra step before getting to the form to
stop spammers) has been added. As have a background image and a little
png of a flower to make it all pretty.
The hack-y banner image from the Elham primary site is now gone as
well, finally.